[추가 / 수정] 보관함 API 페이징 기능 추가 및 즐겨찾기 API 기능 추가
소설, 보관함, 작가 즐겨찾기 추가 및 삭제 기능 개발

+++ NovelForm/src/main/java/NovelForm/NovelForm/domain/box/domain/BoxItem.java
@@ -17,7 +17,7 @@ public class BoxItem extends BaseEntityTime {
     private String status;
 
     @Column
-    private Long is_lead_item;
+    private Integer is_lead_item;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "box_idx")
@@ -29,7 +29,7 @@ public class BoxItem extends BaseEntityTime {
     private Novel novel;
 
 
-    public BoxItem(Long is_lead_item, Novel novel) {
+    public BoxItem(Integer is_lead_item, Novel novel) {
         this.is_lead_item = is_lead_item;
         this.novel = novel;
         this.status = "activated";
